Removed cpu limits for calico-node, calico-typha, and the calico autoscalers (calico-node-vertical-autoscaler, calico-typha-horizontal-autoscaler, calico-typha-vertical-autoscaler) and increased the memory limits.
The memory limits of calico-node and calico-typha will no longer be autoscaler, but they are static at a fairly
high level. CPU limits are no longer enforced to prevent unnecessary throttling.
The downside of the current approach is that it will only be applied to new clusters. Existing clusters will stay
with the existing limits, which can be removed, though.