Reset `NetworkUnavailable` condition after `calico-node` startup.
`calico-node` sets the `NetworkUnavailable` condition to `false` during
startup and to `true` in a `PreStop` hook. This assumes that the
`PreStop` hook always executes before a new instance comes up.
Unfortunately, in some scenarios, this may be a race condition.
Therefore, this change introduces a new side car container, which
executes setting the condition some time after initial startup. It
reuses the `calico-node` startup command, which seem to have no negative
side-effects.