Make source network address translation configurable in IPv6 scenario.
Previously, source network address translation (SNAT) was always
disabled via the `IPPool` configuration in the IPv6 scenario. This
should be a sensible default for scenarios when globally routable IPv6
address space is used. However, if unique local addresses (ULA) are used
it might be useful to still perform source network address translation
on the node level.
This change exposes an option so that SNAT can be enabled for IPv6.