Reduce the size of the cni-plugins container image.
The cni-plugins container used with multus is unnecessarily large and
contains a lot more than required. It actually only needs to include the
cni binaries, a shell, a copy and sleep command.
This change creates a new container image based on the original
including only what is needed on top of an alpine base image.